updated definition of EID to reflect changes in stapl_release/include/graph/BaseGraph.h
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@888 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ OBPRM/BasicDefns.h
@@ -50,7 +50,7 @@
 /**@name General data structures*/
 //@{
 typedef short SID;///< set id type
-typedef short EID;///< element id type
+typedef int EID;///< element id type
 //@}
 
 /// Legal Types of Collision Detecters
